Create new module rank and move scoring functions to rank.
